src/personalization_system.py
```python
from nlp_processing import processPreferences

class PersonalizationSystem:
    def __init__(self, aiCharacter, investorPreferences):
        self.aiCharacter = aiCharacter
        self.investorPreferences = investorPreferences

    def adjustAICharacter(self):
        processedPreferences = processPreferences(self.investorPreferences)
        for preference, value in processedPreferences.items():
            if hasattr(self.aiCharacter, preference):
                setattr(self.aiCharacter, preference, value)

        return self.aiCharacter

def personalizationUpdate(investorPreferences):
    personalizationSystem = PersonalizationSystem(aiCharacter, investorPreferences)
    updatedAICharacter = personalizationSystem.adjustAICharacter()
    return updatedAICharacter
```